+++ src/format.ts
@@ -12,6 +12,7 @@ export type Formatter<T> = (log: Log, options: FormatOptions<T>) => T
 export type FormatOptions<T> = {
   depth?: bigint
   timestamp?: boolean
+  prettyJSON?: boolean
 } & (
   | {
       kind: Exclude<Format, Format.Custom>
@@ -45,15 +46,23 @@ function applyJSONFormat<T>(log: Log, options: FormatOptions<T>): T | string {
   )
 }
 
+function stringifyData<T>(data: any, options: FormatOptions<T>): string {
+  return typeof data === 'object' ? stringify(data, { depth: options.depth ?? -1n, pretty: options.prettyJSON ?? false }) : `${data}`
+}
+
 function applyPLAINTEXTFormat<T>(log: any, options: FormatOptions<T>): T | string {
   const timestamp = options.timestamp ?? false
+  const hasTag = log.tag !== undefined
   let result = `[${LogLevel[log.level].toLowerCase()}]`
   if (timestamp) {
     result += `[${applyDateFormat(new Date())}]`
   }
-  result += `: ${log.message}`
+  if (hasTag) {
+    result += `<${log.tag}>`
+  }
+  result += `: ${stringifyData(log.message, options)}`
   if (log.data) {
-    result += `\n${typeof log.data === 'string' ? log.data : stringify(log.data, { depth: options.depth ?? -1n })}`
+    result += `\n${stringifyData(log.data, options)}`
   }
   return result
 }