+++ src/logger.class.ts
@@ -1,5 +1,5 @@
 import { LogContext } from './log-context.class'
-import { ILogger } from './logger.interface'
+import { ILogger, LoggerOutput } from './logger.interface'
 import { isValidLogLevel, LogLevel } from './log-level'
 import { Format, applyFormat, FormatOptions } from './format'
 import { createVault, IVault, Vault, VaultOptions } from './vault'
@@ -9,7 +9,8 @@ export type LoggerOptions<LogFormat> = {
   format: FormatOptions<LogFormat>
   vault: VaultOptions
 }
-export class Logger<LogFormat> implements ILogger {
+
+export class Logger<LogFormat> implements ILogger<LogFormat> {
 
   private options: LoggerOptions<LogFormat>
   private vault: IVault<LogFormat | string>
@@ -36,41 +37,41 @@ export class Logger<LogFormat> implements ILogger {
    * @param tag The tag to add to meta
    * @returns A log context with the tag
    */
-  tag(tag: string): LogContext<Logger<LogFormat>> {
-    const logContext = new LogContext<Logger<LogFormat>>(this)
+  tag(tag: string): LogContext<LogFormat,Logger<LogFormat>> {
+    const logContext = new LogContext<LogFormat,Logger<LogFormat>>(this)
     return logContext.tag(tag)
   }
 
-  data(data: any): LogContext<Logger<LogFormat>> {
-    const logContext = new LogContext<Logger<LogFormat>>(this)
+  data(data: any): LogContext<LogFormat,Logger<LogFormat>> {
+    const logContext = new LogContext<LogFormat,Logger<LogFormat>>(this)
     return logContext.data(data)
   }
 
-  error(message: any, meta: object = {}): void {
-    this.log(message, LogLevel.Error, meta)
+  error(message: any, meta: object = {}): LoggerOutput<LogFormat> {
+    return this.log(message, LogLevel.Error, meta)
   }
 
-  warn(message: any, meta: object = {}): void {
-    this.log(message, LogLevel.Warn, meta)
+  warn(message: any, meta: object = {}): LoggerOutput<LogFormat> {
+    return this.log(message, LogLevel.Warn, meta)
   }
 
-  info(message: any, meta: object = {}): void {
-    this.log(message, LogLevel.Info, meta)
+  info(message: any, meta: object = {}): LoggerOutput<LogFormat> {
+    return this.log(message, LogLevel.Info, meta)
   }
 
-  debug(message: any, meta: object = {}): void {
-    this.log(message, LogLevel.Debug, meta)
+  debug(message: any, meta: object = {}): LoggerOutput<LogFormat> {
+    return this.log(message, LogLevel.Debug, meta)
   }
 
-  trace(message: any, meta: object = {}): void {
-    this.log(message, LogLevel.Trace, meta)
+  trace(message: any, meta: object = {}): LoggerOutput<LogFormat> {
+    return this.log(message, LogLevel.Trace, meta)
   }
 
-  private log(message: any, level: LogLevel, meta: object = {}): void {
+  private log(message: any, level: LogLevel, meta: object = {}): LoggerOutput<LogFormat> {
     if (!isValidLogLevel(level, this.options.level)) {
-      return
+      return 
     }
     const result = applyFormat({ message, level, ...meta }, this.options.format)
-    this.vault.store(result)
+    return this.vault.store(result)
   }
 }